build!: remove compile flags from the KokkosComm target
Rationale: because KokkosComm is header-only, it is an `INTERFACE`
library in CMake jargon. This means that any targets that link against
KokkosComm would inherit the compile flags/definitions/features/etc that
we had set here. This is often undesirable, and considered bad practice.
In a follow-up PR, we will provide CMake configure/build/test/workflow
presets to allow developpers to run custom builds of KokkosComm for,
e.g., debugging, testing, benchmarking or packaging/distributing
purposes.

Note: since KokkosComm is a C++20 project and relies on some of its
additions, we still enforce the `cxx_std_20` feature (w/o GNU
extensions) so that targets linking against KokkosComm inherit it.