refactor(mpi): move `mpi_type` in the `mpi` namespace
This places the `mpi_type` datatype conversion function (and friends) in
the `mpi` namespace. The rest of the changes is to accomodate any code
that would otherwise fail to compile.

In the future, we may consider rethinking how this is implemented. We
could expose a generic function, templated on the communication space
and specialized as necessary:
```cpp
template <CommunicationSpace CS, typename T>
[[no_discard]] constexpr auto dtype() -> CS::datatype_type;

static_assert(dtype<Mpi, double>() == MPI_DOUBLE);
```

Signed-off-by: Gabriel Dos Santos <gabriel.dossantos@cea.fr>